Home Work 3

+++ HotDrinks.java
@@ -1,8 +1,9 @@
 import java.util.Calendar;
 
 public class HotDrinks extends Beverages {
-    private double temperature;
-    public HotDrinks(String name, int price, int quantity, Calendar bestBefore, int volume, double temperature){
+    private int temperature;
+
+    public HotDrinks(String name, int price, int quantity, Calendar bestBefore, double volume, int temperature){
         super(name, price, quantity, bestBefore, volume);
         this.temperature = temperature;
     }
@@ -11,7 +12,7 @@ public class HotDrinks extends Beverages {
         return temperature;
     }
 
-    public void setTemperature(double temperature) {
+    public void setTemperature(int temperature) {
         this.temperature = temperature;
     }
